requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
ultralytics==8.0.196
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.0.76 